kinda changed my mind about multiple files moved things around setup for using discord embed data modifications for bopae piece image url
