+++ bopae/bopaecmp.py
@@ -1,87 +1,5 @@
 import os, re, json
 
-class Bopae:
-    """BNS soul shield utils"""
-
-    def __init__(self, data_dir : str):
-        self.bopaeData = {}
-        self.data_dir = data_dir
-        self.reload()
-
-    # Reloads soul-shield data
-    def reload(self):
-        self.bopaeData = {}
-        for file in os.listdir(self.data_dir):
-            with open(self.data_dir + file) as f:
-                self.bopaeData.update(json.load(f))
-
-
-    # Lists all soul-shield sets in the database
-    # return : str
-    def list(self):
-        """Show SS sets in the database"""
-
-        multiline = "Available SS sets:\n{}\n".format(", ".join(map(str, self.bopaeData)))
-        # multiline += "Available stats: AP cRate cDmg PEN aDmg Ele CCdmg ACC "
-        # multiline += "HP1 HP2 DEF cDef VIT REG EVA BLK rDmg fusionmax"
-        return multiline
-
-
-    # soul-shield search.
-    # message : str
-    # return : str
-    def search(self, message : str):
-        """BNS soul-shield search.
-
-        FORMAT:
-            !bopae ([set name] [slot num/all]...)...
-        EXAMPLES:
-            !bopae tomb 1
-            !bopae yeti 2 4 ebon all asura 1 3 5
-        """
-
-        query = message.split()
-        if query[0].lower() == "search":
-            query = query[1::]
-        query = self._parser(message.split()[1::])
-
-        multiline = query["multiline"] + "\n"
-        del query["multiline"]
-
-        for bopaeset in query:
-            # try:
-                multiline += ("# # # # # # # # # #\n"
-                            "Set name: {} [{}]\n"
-                            "Notes: {}\n"
-                            "\n").format(self.bopaeData[bopaeset]["setName"],
-                            bopaeset, self.bopaeData[bopaeset]["setNotes"])
-
-                multiline += "Set bonus:\n"
-                for i in sorted(self.bopaeData[bopaeset]["setBonus"]):
-                    multiline += "{} set: {}\n".format(i, self.bopaeData[bopaeset]["setBonus"][i])
-                multiline += "\n"
-
-                for reqSlot in query[bopaeset]:
-                    reqBopae = self.bopaeData[bopaeset]["slot"+str(reqSlot)]
-                    multiline += (
-                        "Slot {}\n"
-                        "HP1: {}\n"
-                        "Fusion max: {}\n"
-                        "Primary stat: {} {}\n"
-                        "Secondary stats: {} {}\n"
-                        "\n"
-                        ).format(reqSlot,
-                        reqBopae["HP1"], reqBopae["fusionmax"],
-                        reqBopae["stat1"], reqBopae["data1"],
-                        reqBopae["stat2"], reqBopae["data2"]
-                        )
-
-                multiline += "\n"
-            # except:
-            #     multiline += "Internal error\n"
-        return multiline
-
-
     # soul-shield side-by-side listing
     # format: !bopae compare [name1] [name2] [slotnum]
     # message : str
@@ -227,44 +145,6 @@ class Bopae:
                     multiline += '{}{}: {}\n'.format(' '*29, i, piece2['data2'][0:2])
 
         return multiline
-
-
-    # TODO better exception handling
-    # takes in text:list. returns dictionary key:setName, value:list integer slots
-    # multiline in dict: multiline string of errors. "" on empty
-    def _parser(self, text):
-        currset = ()
-        query = {"multiline": ""}
-        for i in text:
-            name = self._namesearch(i)
-            if name == "":
-                if i == "all" and currset != ():
-                    query[currset] = [1, 2, 3, 4, 5, 6, 7, 8]
-                else:
-                    try:
-                        i = int(i)
-
-                        if currset == ():
-                            raise Exception('Attempted to assign soulshield slot to empty set')
-                        if i < 1 or i > 8:
-                            raise Exception('Invalid soul shield slot')
-
-                        if currset not in query:
-                            query[currset] = list()
-                            query[currset].append(i)
-                        elif i not in query[currset]:
-                            query[currset].append(i)
-                    except:
-                        if currset == ():
-                            query["multiline"] += "Invalid set name [{}]\n".format(i)
-                        else:
-                            query["multiline"] += "Invalid slot num [{}] for set [{}]\n".format(i, currset)
-            else:
-                if name not in query:
-                    query[name] = list()
-                currset = name
-
-        return query
 
 
     def _namesearch(self, query):
