--- libbopae/statistics.py
@@ -1,15 +1,0 @@
-
-
-class Bopae_statistics:
-    # def bopae_statsearch(self, query):
-    #     # TODO optimise
-    #     # " AP cRate cDmg PEN aDmg Ele CCdmg ACC HP1 HP2 DEF cDef VIT REG EVA BLK rDmg fusionmax "
-    #     if len(query) < 2:
-    #         return ""
-    #
-    #     query = re.compile(query.lower())
-    #     for i in ["AP", "cRate", "cDmg", "PEN", "aDmg", "Ele", "CCdmg", "ACC", "HP1", "HP2", "DEF", "cDef", "VIT", "REG", "EVA", "BLK", "rDmg", "fusionmax"]:
-    #         if query.search(i.lower()) != None:
-    #             return i
-    #
-    #     return ""
